civicai/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pandas==2.2.3
numpy==2.2.1
faiss-cpu==1.9.0.post1
transformers==4.47.1
torch==2.5.1
sentence-transformers==3.3.1
langdetect==1.0.9
googletrans==4.0.0-rc1
scikit-learn==1.6.0
python-dotenv==1.0.1
slowapi==0.1.9
email-validator==2.2.0

Pillow==11.0.0
